requirements.txt
fastapi>=0.110
uvicorn>=0.23
scikit-learn>=1.3
numpy>=1.26
pillow>=10.0
pydantic>=2.0
